Add file to auto-generate graph objects.
Previously, we generated these *on import*, which is slow and it’s a
little sneaky. Users end up using classes that are not actually written
*anywhere* in our package.

Now that we only update the plot schema in version releases (no longer
on import), we can also just programmatically add in all the graph obj
classes. Instead of adding them to memory, we literally just print them
to the file.

This is some serious meta-programming… but, the resulting file will have
to be syntactically valid and pass our test suite, so I’m not *too*
worried about it.