Fix duplicate import source `graph_reference`.
In Python 2, you *must* have an `__init__.py` file to have a python
package. However, in Python 3, you don’t need that!

We have `/graph_reference` and `graph_reference.py` defined, which seems
to *work*, but is poorly organized and should be fixed. For instance, it
causes `unresolved import errors` in IDEs that have `goto-source`-type
functionality.

Conflicts:
	makefile